components/model-training/pyproject.toml
[project]
name = "model-training"
version = "0.1.0"
description = "LoRA training modules for MedGemma"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.45.0",
    "peft>=0.10.0",
    "bitsandbytes>=0.43.0",
    "accelerate>=0.30.0",
    "datasets>=2.19.0",
    "torch>=2.0.0",
    "scikit-learn>=1.0.0",
    "fsspec>=2023.10.0",
    "google-cloud-aiplatform>=1.130.0",
    "google-cloud-storage>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = ["pytest"]

[tool.hatch.build.targets.wheel]
packages = ["src/model_training"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
